test(core): split selector parser specifications
Split the 328-line parser spec into focused behavioural suites: model
contracts (EntitySelectorModelTest) and one parsing suite per grammar
family (scalar, filter, map, SNBT), keeping only selector-level
orchestration in EntitySelectorParserTest. Assertions move verbatim; a
shared assertParseFailure helper replaces per-file copies.

The direct SelectorReader.readQuotedString() test is removed: it
asserted an internal seam, and the equivalent public-surface failure
stays pinned via @e[name="bad\q"].
